cranky: Add new script cranky-dput-sources
This is a helper script for uploading kernel source packages to a PPA.
It works similarly to cranky-build-sources in that it cd's to the
different source directories that make up a set of kernel packages,
constructs the source package name from the changelog, checks that it
exists in the parent directory, displays its name and asks for
confirmation before doing the actual upload.

The 'URL' of the upload PPA is gathered from the output of
cranky-list-routing.

Acked-by: Connor Kuehl <connor.kuehl@canonical.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>