cranky/docs: move "cranky tag" after "cranky test-build"
If "cranky test-build" fails and we need to commit additional changes we
also need to re-create the tag, otherwise the tag would point to an old
commit instead of the HEAD.

To prevent this problem from happening move the "cranky tag" step after
"cranky test-build".

Also add a note to explicitly re-run "cranky close" if a test build
fails and additional commits need to be added.

Signed-off-by: Connor Kuehl <connor.kuehl@canonical.com>
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>