cranky/test-build: always delete remote branch
The gbk hook is only triggered when the remote branch is updated. For
the test build we do a force push to trigger the build and delete the
remote branch when the build finishes. If we cancel the build in the
meantime the remote branch is not deleted and trying to build the same
sha1 again will only work if the remote branch is deleted.

Always delete the remote branch before a build to avoid having to do it
manually in such cases.

Acked-by: Juerg Haefliger <juergh@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>