cranky/cranky-checkout: Always (re-)configure the git repos
When a release transitions to ESM, the kernel mailing list for patches
changes. And we might also add additional configs in the future. So run
the config step everytime and not just after a repo is freshly cloned.

While at it, move the addition of the security repo to git_configure()
so that the config steps are all grouped together and fix pylint3
indentation errors.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>