cranky: checkout: Use remote as part of the local branch name
We already have some logic implemented in the remote name generation
in order to be it as brief as possible and yet unique considering the
user's specific configuration.

Use the remote name as part of the local branch name to make sure the
branch is also unique for any kind of configuration (for example, when
using a single repo for all the kernels).

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>