kernel-stable-dashboard: Not Releasable only applies when testing was complete
We are only going to be marking debs as Not Releasable if testing is
complete but the kernel has not moved on to promoting-to-updates etc.
In that case we try and find the cause and if we cannot we assume it is
something not-knowable and report the generic Nor Releasable.  This however
should only be done if we have completed testing.  Otherwise we will
incorrectly report deb status of Not Releasable for non-debs variants
such as snap-debs.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>