ktl/tracking_bug2.py: Optimize tag creation
We really want to set the 'kernel-release-tracking-bug' tag as
soon as possible (at least before adding the non workflow task)
because this stops the kernel bot from asking for logs (which
turns out to be one of the reasons for failing to create the
tracking bugs).

While doing that it felt really bad to have the tag reset function
first delete all tags and then add back the initial set. That would
open a gap for the kernel bot. So instead, this is changed to only
drop those tags which would not be added and then add only those
not already present later.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>