kernel-stable-dashboard: use swm status.yaml for bug listings
The existing data source kernel-sru-workflow.json is very expensive to
generate and therefore only rebuild infrequently.  Switch over to using
the SWM status file as, while also expensive, we are already necessarily
keeping this up to date.  This (now) has all of the required
information.  Switching over also gives us a more consistent view as it
now is based on SWMs view of the world,.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>