swm: swm-trello-sync -- trello card cranker
Add a new swm-trello-sync handler which monitors the swm reported status
for the various kernel prepare-package and snap-promote-to-edge tasks.
It uses this information to determine the appropriate column to which
the associated card should be moved.  When the card is not in the right
place it is shifted to the appropriate list.  Some amount of manual
leading is permitted, for instance if the card should be 'ready to crank'
it would move to 'On Deck' but if you manually pull it to 'In Progress'
it is deemed acceptable and not moved back.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>