swm: bug -- populate a proper per bug target
A tracking bug is associated with two things a source package (specifically
the kernel-series source object) and either a set of debian packages or
a snap.  Currently the we expose the former using the series and source
name pairing.  We currently expose the snap-name for a snap-debs variant
via a target indicator on the task.  Generalise this to an overall target
for the tracking bug.  For a debs variant this will be the name of the
main package (typically the same as the name of the kernel-series source
object).  For a snap-debs variant it will be the snap-name.  The overall
target of the tracker then becomes series:source:target in all cases.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>