cve-tools: Ignore commit info for non-kernel packages
Ignore CVE break and/or fix commit info when encountering a source
package that has a name starting with the "linux-" prefix but does not
produce a kernel.

For example, it is incorrect to consider a commit ID that fixes a CVE in
the linux-firmware package to be a valid kernel commit ID. A single CVE
file in the Ubuntu CVE Tracker could contain entries for the
linux-firmware and linux source packages when a firmware update and
kernel code changes are both needed to address a CVE.

A blacklist of non-kernel source packages is not an ideal filter as it
is not future proof. However, new kernel source packages are added to
the archive more often than new userspace source packages that contain
the "linux-" prefix so a blacklist is easier to maintain.

Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Tyler Hicks <tyhicks@canonical.com>