cranky: crl/handle -- Make 'base-path' a top-level config option
The option 'base-path' can/will be used for other purposes as well, so
make it a top-level option rather than a sub-option of 'package-path'.

For the time being, also allow the deprecated option and emit a warning
to inform the user that the format of the .cranky config file has changed.

Acked-by: Connor Kuehl <connor.kuehl@canonical.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>