gen-workflow-report: handle bad titles better
Split will only return one element when there is no separator found.
Attempting to push this into a two part tuple will raise and Exception.
As we are only interested in the first element simply ask for the first
one only.  This will leave us with an invalid package, but in that context
we will later reject the entry.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>