swm: wft/workflow -- if we are closing phase should be Complete
When we finally are ready to close out a tracking bug we currently just
leave the phase alone on the assumption the record will be dropped at the
end of the run.  But when we reach there via a manual direct bug scan we
do not remove it (though likely we could and should).  At the point we
close it mark the phase as Complete in case it is reported before
removal.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>