swm: bug,wft/workflow -- keep the top level bug open after completion
Tooling cares about the final tracking bug status being reported in
status.yaml.  For example we want to move completed trello cards to Done
on the final action.  Humans also like to see that the final action was
complete in the dashboard.  Hold the workflow task open for an hour
after the bug moves to phase Complete.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>