maint-modify-patch: delete temp file on errors
If an error happens while the patch file is processed, e.g. a key error
caused by a non-existing irc alias, the script breaks but the temporary
file is not removed. Fix it by removing the file when an exception
happens.

v2: merged both 'with' clauses to avoid a massive reintentation.

Acked-by: Connor Kuehl <connor.kuehl@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>