swm: swm-publishing -- initial modification monitor
We currently scan all live bugs every 30m.  When we have an old cycle
in a dwell between -updates and -security, and a new cycle is being
started with trackers created we can have over 100 trackers in play.
This takes some 20m to scan so we are effectivly scanning 80% of the
time.  This is highly wasteful as most are unchanged in any particular
run.

A tracker only needs to be scanned by swm if it has changed, or if
something it is waiting for has changed.  These include publications of
sources and builds in various archives, publication of snaps, and testing
results coming in.

Put together a monitoring framework.  This is intended to be run at a high
frequency and when nothing is changed to complete very quickly.  For each
monitored item we use any 'only tell me what has changed' interfaces
supplied.  This is designed to identify any changed trackers, apply swm to
those, and then quit.  Any state required to allow continuation is stored
in the monitor persistent store (which is backed to swm-publishing.yaml);
it is only synced to disk on a successful swm run over the triggered
trackers.

In testing with a full bug load and all the monitors a full scan takes
less than 30s.  This allows us to sensibly cron the publishing run at */5.
Meaning trackers, swm-ls, and the dashboards are much fresher always.

This patch adds the framework and persistant monitor backing store
framework.  Future patches add the individual monitors.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>