swm: snap -- switch up to /v2/snaps/info API interface
Switch from the deprecated /v1/snaps/details/ to /v2/snaps/info/ interface.
This allows us to interrogate all of the channel publications for a snap at
the same time reducing round trips to the store.  Also update the internal
form to a dict to allow easier addition of new fields going forward.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Connor Kuehl <connor.kuehl@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>