stable/create-kernel-tasks: Use spin number when searching
If a spin number is provided on the command-line, then this is
used to create a list of cycle tags to search for instead of
the tracking bug live tag.

This should allow to create re-spins for source after the parent
tasks were already closed. However, because this also will find
any closed tasks of the source we want to re-spin, this needs
some additional check to prevent duplicating a tasks which
ran to completion against a re-spin task.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>