swm: snap -- expose released-at field in the channel_map and make it public
We are going to care about the publishing date in each channel so add
the released-at field to the requested data list and to the results.  We
will also want to obtain the full channel publication map to allow
iteration over it, expose the currently private channel_map as a public
method.  This also allows repeated calls avoiding the internal caching
which is not necesarily suitable in a public API.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>