ktl/tracking_bug2.py: Make snapd-debs loadable
Right now, the optimized load function misses all snap-debs because it
relies on a package task to make the series filter work. But that task
does not exist with snapd-debs tracking bugs.

Luckily those have the series in the title and that also can be accessed
from the task list which is returned by searchTasks. So we can quickly
filter without having to first create a TrackingBug object.

Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>