stable/create-kernel-tasks: Duplicate snap tasks
The duplication of snap tasks did not work since they were not loaded
properly but also because they need special care when it comes to
naming.

The following fix changes loading to include all types of tasks and
ensures to use a special form of <source name>/<snap name> when
looking for existing snap-debs tracking bugs.

Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>