cranky/docs: add `verify-release-ready` step
This script can help notify a crank turner of certain mistakes that
will likely result in a revision before the kernel passes review for
upload. By adding this step here, those mistakes can be corrected
before source packages are built and the kernel is reviewed. This
will ultimately save time for both the reviewer and the crank turner.

Signed-off-by: Connor Kuehl <connor.kuehl@canonical.com>
Acked-by: Khalid Elmously <khalid.elmously@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>