link-to-tracker: support incoming series/ prefix on tracker titles
We are slowly adding series/ prefixes on the tracker titles.  This
makes it easier for humans and computers alike to know what a tracker
is referring to.  This exposes a bug in matching caused by the
introduction of snap-debs variants.  We currently are ignoring those
because they have the new series/ prefix, once we accept those we find
duplicates.  Add additional validation to only include '-proposed
tracker' bugs.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>