swm: manager -- reset bug reason data on each crank pass
We currently reset reason data when we first touch a bug.  This leads us to
accumulate reason data for a bug over successive crank runs.  For example
if promote-to-proposed task was in New (because its bug was reset) but the
packages are fully in -proposed, then the task will be cranked three times
(New -> Fix Committed -> Fix Released).  If one of the early cranks sets
a reason but the last need not then the outdated reason will remain.

Reason data should be reset on each re-crank iteration.  Move the reset
into the loop.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>