cranky: close helper: add option to provide a base version
For backports of derivatives, the 'close' in-tree helper uses the
previous backport changelog version to construct the base version
provided to 'insert-ubuntu-changes' in order to get the new changelog
entries from the base derivative changelog. Add a '-b' option to force
the script to use the provided base version. This is useful when the
base changelog is changed for some reason, e.g. when updating to the
derivative of a newer series.

This is the case when we move an edge which is based on one derivate to
a derivative in a newer series. In that case the rebase replaces the
master changelog by a version which does not contain the version which
the edge kernel previously was based on. And unlike the -v of
dpkg-buildpackage this results in selecting nothing at all.

To work around this we need to provide an override to the previous base
version on the close step.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>