swm: git_tag -- switch to using git ls-remote to handle private repositories
Currently we can only verify data for public git repositories.  Switch
to using git ls-remote <respository> and parsing the returned refs.
By careful configuration of git (see below) we can ensure that this works
for both public and private repositories.

This requires the calling user to have an ssh key in launchpad and for
the git prefixes required to be configured to use ssh and the appropriate
launchpad account in .gitconfig, something like:

    [url "git+ssh://ubuntu-kernel-bot@git.launchpad.net/"]
        insteadof = lp:
        insteadof = git://git.launchpad.net/
        insteadof = https://git.launchpad.net/

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>