swm: bug,manager -- add refresh_time accumulator
Some input triggers are time related.  For example we wait 1 hour after
publishing to -proposed before triggering testing, this allows time for
the publisher to publish, and mirrors to replicate the new packages.
We also wait 24 hours after publishing to -updates before triggering
promote-to-security.

In order to trigger scanning of bugs based on these kinds of events we
need to have an idea of when this bug would like to be checked next.
Do this by recording the first triggering time as a refresh_time in the
the status.yaml so that swm-publishing can read and honour it.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>