adt-matrix: ignore results triggered by non-kernel packages
This reverts commit 7afe623cfac6c7844b8b021877fe95f7cab1d4e2,
which commented out some code to ignore results when testing was
not triggered by the kernel package. Now non-kernel triggers will
result in only the rebuild test running, and not the regression
tests. For promoting kernels we only want results where the full
regression test suite has been run, so uncomment this code.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>