ktl/tracking_bug2.py: Fix pkg nomination status
The old code did set all states of the package tasks (including
the normination) which was lost when changing to add those after
all workflow tasks were added.
Fix this by searching for the approved nomination for the series
when it is done.

While one it, the whole checking for new package names that was
copied from the old implementation seems to be rather pointless
now. So remove it.

And also, there was a whole block only being done if a launchpad
bug had been created. However, things were reworked in a fashion
that execution should never get that far without. So this test
can go away as well.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>