cranky: git-ubuntu-log -- support new Properties: changelog tag
We currently have Ignore: yes as a flag to indicate we should
pretend this commit is missing.  We have a second category
of commits which we want to include in the changelog but not
trigger test builds on.  For these we need a new flag, to prevent
a proliferation of arbitrary flags add a new Properties: field
and let that implement 'ignore' and 'no-changelog' as meaning
do not include this in they changelog.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>