mainline-builds: cod-changeset -- add support for rebase trees
We are wanting to scan rebase trees and build changesets for the new tips.
Currently when a tip is non-linear we simply test the new tip and hope
that is sufficient.

For this we need to work out which of the commits on the new tip were
really on the old tip and elide them from the changeset list.  For
example:

    ---A---B---C
        \       `-M1--M2--M3
         `-M1--M2

In the graph above we have both rebased against a master branch from A
to C and added a new commit M3.  We would expect the changeset detector
to note that M1 and M2 are the same and therefore not interesting.
M3 on the other hand is shiney and new and therefore is interesting.
The changes from A to C are not individually interesting either but
a rebuild of M2 is appropriate to catch where the rebase component is
triggering a build issue.

To determine what is new on the new C based tip we use the git patch-id
for each change on the old tip and scan backwards on the new tip looking
for a change from the old tip, when we find one we know we are highly
likely to be at the start of the 'moved' content.

Note that this might mean we miss some new change if you do serious
reordering surgery on it say inserting a new changeset under the top set.
But this is not something we routinely should be doing and missing those
changeset is not likely to be hugely problematic.

Note that this also adds support for Properties: no-test-build to allow us
to avoid rebuilds of classically pointless commits such as link-to-tracker
and cranky-tooling updates.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>