kernel-stable-dashboard: handle swm status.yaml containing dates
When there is a pending refresh-time: on a record this is not representable
in JSON.  This triggers the dashboard to fail to update until the
refresh-time expires and is removed.  Add a JSON encoder to allow this
to be represented.  At the same time pull the open coded python into a
proper script.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>