mainline-builds: cod-changeset-build -- add support for an order field
When a tree has been rebased the commit timestamps can be identicle
for a number of sequential commits.  This makes it impossible to ensure
the order in the reports is the same as the tree.  We are going to add
an order idenfier which will be higher for later commits.  This will be
combined with the commit time to ensure they order correctly.  If we are
passed such an order add it to the commit time in the result.

Signed-off-by: Andy Whitcroft <apw@canonical.com>