mainline-builds: cod-changeset-trigger,git-changeset -- add support for an order field
When a tree has been rebased the commit timestamps can be identicle for
a number of sequential commits.  This makes it impossible to ensure the
order in the reports is the same as the tree.  Add an ordering element.
This is a number which is higher for later commits.  This will be combined
with the commit time to ensure they order correctly.  Calculate and pass
the order constraint to the backend.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>