swm: manager,bug -- add live tracker links to a tracker as we scan it
Use the current live state information to accumulate a list of trackers
which have this scanned bug as its master-bug link.  Add these to the
swm-properties under a tracker: key.  Record all live bugs under the
series/source/target key, replacing the existing list for that key.
This should normally be a single bug.  We only add keys, so that now
closed trackers are not lost.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>