swm: wft/promote-signing-to-proposed -- only report source status in ppa
Generally a copy task is idle till its input is ready for copy.  Once the
copy is ongoing we monitor the destination and report any issues in
that destination.  However for promote-to-proposed (only) we care also
about the source and report it; we do this because there is no other
task monitoring the builds in progress.  As things stand we end up double
reporting issues as destination issues for promote-to-proposed and source
issues for promote-signing-to-proposed.  Ensure we do not report source
status for promote-signing-to-proposed instances.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>