cranky/cranky-update-snap: Handle dual source snaps
With recent changes to GKE kernels we introduced the concept that a snap
can have the same name but being produced by different sources and
getting released to different channels (track/risk).

This modifies the update script for snaps to extract the source name
from the parameter section for the snap build using an additional
parameter (KERNEL_SOURCE) which is added solely for that purpose.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>