stable/sru-cycle-kernel-versions: Handle tb title adjustments
Workflow was changed to include the series in the title string. This
causes sru-kernel-versions to no longer find any master kernels. To fix
this, add the optional element to the regex.

And while at it replace the group index numbers by named elements which
make it much simpler to access them later. Also break up some long
lines.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>