link-to-tracker: add a --source/--version pair
When kernels are prepared in advance of the creation of the tracking
bugs it is desirable to attach a specific version to a tracking bug
for a specific source package without reference to a local tree.

Add a --source and --version options to allow specification of the source
and version mapping.  The tracker is looked up as normal and the version
applied.  No reference is made to the local tree and no data is written
back to it.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>