ktl/tracking_bug2.py: Include invalid tasks when loading
When loading a set of tracking bugs we search all tasks which
satisfy a certain status and tag list criteria. To support
filtering by series we have to look at the series nomination
of the package task of debs tracking bugs (or combo). But this
is set to invalid for sources which have not been uploaded to
the archive (,yet).

To make duplication work for those, we also have to include
invalid tasks when searching.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>