ktl/tracking_bug2.py: Add support for hard invalidate
Some tracking bugs will need to be invalidated in a way that we will
not find them ever again. This is done by removing the cycle tag
together with the live tag before invalidating the tasks.

This adds an option which can be used by external callers. The default
remains as it was before.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>