swm: wft/prepare-package -- handle direct copies to -proposed better
When we decloak a CRD but which will have tracking bugs we end up with the
prepare-package tasks remaining in Confirmed untill everything is complete.
This leads the dashboard to report 'ready to crank'.

We already track the state of this task through its various stages
closing only once uploaded and tags pushed.  Add state handling such
that we automatically move In Progress as soon as we have a version,
and Fix Committed once uploaded.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>