swm: swm-trello-sync -- reimplement in python3
We are trying to do something which is very complex in shell which is
wholey inadequate for the task.  In trying to identify lost entries we
are making an unmitigated mess in the face of duplicates and we are unable
to rectify it sensibly.  Reimplement this thing in python3 as all of our
data sources are python libraries anyhow.

A few things got a lot better in the conversion.  For example we now can
detect and handle duplicate cards sensibly and move them to lost+found.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>