cranky: master -- getabis: skip up-to-date check when removing old abis
When the zfs build is disabled, our packaging automatically adds
the zfs modules to <oldabi>/modules.ignore. Usually this file
doesn't exist, but if it does (i.e. it is known to git) the rm
of the old abis directory will fail because this file has local
modifications. Use the -f flag for git-rm to override the
up-to-date check and force removal of the old abis.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>