pika: populate shared pika library with latest upstream
We have two existing pika versions in the tree.  One is very old and is
mostly python2 compatible, another is newer and mostly python3
compatible.  The python2 version is within the ktl directory which by an
accident of sematics found by python2 but not at all by python3.  We
will sort all of the references to this library to use this one.

Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>