ktl/tracking_bug2.py: Stop maintaining derivatives
In the old tracking bugs world derivatives and backports were added to
the description as a reference for human consumption. The initial new
implementation of the tracking bug class was keeping that up to date
when adding new derivatives or backports to a master bug. This was ok as
long as this only was done for completely new sets of tracking bugs or
while the SWM scanner was guaranteed to sleep.

However the sleeping time gets shorter and when SWM wakes up, it is
rather quick at finding tracking bugs which have changed since it went
to sleep. And the way create-kernel-tasks works an existing master bug
would be updated a lot while all the derivative tracking bugs of a
re-spin were created. Which prompted SWM to update the master bug and by
that invalidating the copy which create-kernel-tasks had. And that leads
to the dreaded precondition failed explosions.

Since this battle cannot be won without any way of securly locking
access to a tracking bug (which does not exist), change the tracking bug
library to no longer look at or maintain the derivative/backport list in
the tracking bug itself. Instead only maintain that information
dynamically. SWM will actually add the forward links when scanning. So
if that information exists it is used instead. But adding or removing
derivatives from a tracking bug via the library will not change the SWM
properties. Only the master bug reference in the derivative gets set or
unset.

To make up for that a little, I added code which will, when adding a new
tracking bug to a tracking bugs set, automatically be adding any new master
bug to the set and always add the derivative to its master.

Note, while the key of the derivative_bug_ids dictionary is kept being
bug IDs, the value is changed to be <series>/<source>[/<snap>] as it is
in the trackers property. I believe there was no users of the content
externally. Only when updating the description this was used, but that
is gone now.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>