tb-task/tb-task-invalidate: Do not iterate over bug set
The current implementation relied on a set of tracking bugs only
containing those tracking bugs which were actively added. But in future
the set might automatically gain the master bug of an added bug.

So instead of using the bugs in the set, update the list provided on the
command line if a bug number needs to be skipped.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>