swm: wft/workflow -- santise state descriptions and order
We are trying to fit the snap related tasks into the same time order as
debs and they are not really compatible.  In particular snaps have two
separate testing phases which are currently all mixed up.  Split the two
sequences as we actually do not see both in the normal world.  Clean up
the textual descriptions of each Hold so we can more correctly tell what we
are holding to do.  Now separate order the snap tasks in the order they are
expected so we can tell edge and beta apart and see both testing phases.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>