swm: manager -- check for bug duplication as we crank
When cranking a full bug set we do a search for the currently valid bugs
eliminating any duplicates as we go.  We then scan that list linearly
cranking them.  With our current bug load this can take nearly 30m.  In the
window between making the list and cranking the bug it is entirely possible
for the bug to be duplicated leading us to crank a now duplicated bug.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>