stable/verify-meta-versions: New helper script
When we do backports of derivatives (where the package name is the
same), we need to ensure that the version of the meta package in later
series remains higher than in older series. Otherwise upgrades may not
behave as expected.

This script will check all meta package versions of such backports and
report incorrect versions and also gives a warning if the older series
version gets close to the later series.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>