swm: package -- use the official reference to refer to an archive
We currently emit the archive api URL for an archive.  This is neither
human friendly nor directly useful.  An archive has an official name,
the archive.reference which is both useful to look it up and also in a
form humans are expecting.  Switch to this in the various debugging output.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>