swm: swm-publishing -- monitor private launchpad archives
Add MonitorLaunchpadArchivePrivate a monitor for private Launchpad
archives.  When an archive is private the normal getPublishedSources and
getPublishedBinaries combination do not work.  Therefore we switch to
monitoring the pool.  We monitor each of the live series InRelease file
for modification.  When they change we then iterate over all packages
which could be published in this archive checking their current
publication dates.  These dates are accumulated in the monitor
backing store.  When they change the source is triggered as for public
archives.

Plumb this into the launchpad archive factory.  We first try and
instantiate a MonitorLaunchpadArchive on an archive, if it is private
we will get permissions issues.  When these occur we fall back to
MonitorLaunchpadArchivePrivate.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>