swm: swm-publishing -- monitor launchpad projects
Add MonitorLaunchpadProject a monitor for Launchpad bugs in a project.
We monitor tasks using searchTasks with a modified_since date.  We maintain
the latest date in the monitor backing store.  When a bug is modified
we run the activity report for the bug for the period in question.
If after eliding updates by ubuntu-kernel-bot modifications remain we
trigger that bug.

Add a factory function to create these and instantiate one for
kernel-sru-workflow.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>