swm: swm-publishing -- monitor snap publications
Add MonitorSnapStore a monitor for Snaps within a SnapStore.  This pulls
the current publication matrix for the snap and accumulates the latest
publication time for the tracks for this series/snap.  This time is
recorded in the monitor backing store.  When the publication time
changes we trigger this series/source/snap.

Add a factory to add a MonitorSnapStore for each snap in
kernel-series.yaml.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>