swm: swm-publishing -- monitor swm refresh-time records
Add MonitorStatusRefreshTime a monitor for swm refresh-time timestamps.
Some potential triggers are not directly monitorable.  For example there
is modification time of git tags; if they are missing we have to poll
for them to appear.  Also some items are directly time related such
as the tracker close dwell time, and the publish mirror time dwell.
When swm starts one of these dwells it will record a refresh-time for
that tracker in status.yaml.  Sweep status.yaml for these times and when
they are expired trigger the trackers.

Add a factory function to instantiate MonitorStatusRefreshTime on swm's
primary status.yaml file.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>