swm: swm-publishing -- mointor automated testing results
Add MonitorAutomatedTesting a monitor for automated testing (ADT) results.
Watch the status of each series/package for change.  Triggering any source
which has such a change.  The previous status is recorded in the monitor
backing store.

Add a factory function to instantiate a MonitorAutomatedTesting over the
testing results from ADT.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>