swm: swm-publishing -- monitor launchpad queues
Add MonitorLaunchpadQueues a monitor for launchpad queue entries.  When
a source is copied to the primary archive it can jam in the Unapproved
queue for the series if it has signing bits, and later it can also jam
in the New queue.  Monitor for entries in these queues for packages
which are part of sources in kernel-series.yaml.  New entries trigger
change requests so that the reasons are updated to include the queue
status, we also Announce these so they end up more visible.

Add a factory function to instantiate MonitorLaunchpadQueues.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>